Create new PremiumListUtils class and refactor out existing method
