+++ core/src/main/java/google/registry/tools/server/UpdatePremiumListAction.java
@@ -17,6 +17,7 @@ package google.registry.tools.server;
 import static com.google.common.base.Preconditions.checkArgument;
 import static google.registry.model.registry.label.PremiumListUtils.savePremiumListAndEntries;
 import static google.registry.request.Action.Method.POST;
+import static google.registry.schema.tld.PremiumListUtils.parseToPremiumList;
 
 import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableMap;
